+++ pysui/sui/sui_txn/async_transaction.py
@@ -689,13 +689,16 @@ class SuiTransactionAsync(_SuiTransactionBase):
         version="0.20.0",
         reason="Removed recipient. Transfer of UpgradeCap up to user as per Sui best practice.",
     )
+    @versionchanged(
+        version="0.36.0",
+        reason="Drop legacy_digest. No longer supported by Sui.",
+    )
     async def publish(
         self,
         *,
         project_path: str,
         with_unpublished_dependencies: bool = False,
         skip_fetch_latest_git_deps: bool = False,
-        legacy_digest: bool = False,
     ) -> bcs.Argument:
         """publish Creates a publish command.
 
@@ -715,7 +718,6 @@ class SuiTransactionAsync(_SuiTransactionBase):
             project_path,
             with_unpublished_dependencies,
             skip_fetch_latest_git_deps,
-            legacy_digest,
         )
         return self.builder.publish(modules, dependencies)
 
@@ -742,6 +744,10 @@ class SuiTransactionAsync(_SuiTransactionBase):
         version="0.20.0",
         reason="Added package compiled digest control as per Sui 1.0.0.",
     )
+    @versionchanged(
+        version="0.36.0",
+        reason="Drop legacy_digest. No longer supported by Sui.",
+    )
     async def publish_upgrade(
         self,
         *,
@@ -750,7 +756,6 @@ class SuiTransactionAsync(_SuiTransactionBase):
         upgrade_cap: Union[str, ObjectID, ObjectRead],
         with_unpublished_dependencies: bool = False,
         skip_fetch_latest_git_deps: bool = False,
-        legacy_digest: bool = False,
     ) -> bcs.Argument:
         """publish_upgrade Authorize, publish and commit upgrade of package.
 
@@ -779,7 +784,6 @@ class SuiTransactionAsync(_SuiTransactionBase):
             project_path,
             with_unpublished_dependencies,
             skip_fetch_latest_git_deps,
-            legacy_digest,
         )
         # Verify get/upgrade cap details
         if not isinstance(upgrade_cap, ObjectRead):