+++ samples/cmds.py
@@ -37,7 +37,6 @@ from pysui.sui.sui_builders.get_builders import (
     GetTx,
 )
 
-from pysui.sui.sui_utils import publish_build
 from pysui.sui.sui_excepts import (
     SuiMiisingBuildFolder,
     SuiPackageBuildFail,